+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 # Changelog
 
+## 0.4.0
+
+- Added convenience method to `ShaOne` to hash a string
+- Changed `ShaOne.hashData` to `ShaOne.hashBytes`
+
 ## 0.3.0
 
-- Added convenience methods to `SHA` to hash a file from its path
+- Added convenience methods to `ShaOne` to hash a file from its path
 - Added tests
 
 ## 0.2.0
 
-- Changed API to use a static reference to `SHA1`
+- Changed API to use a static reference to `ShaOne`
 
 ## 0.1.0